backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
anthropic==0.52.*
pandas==2.2.*
scipy==1.15.*
pywebpush==2.0.*
apscheduler==3.10.*
python-multipart==0.0.*
httpx==0.28.*
python-dotenv==1.1.*
pytest==8.3.*
pytest-asyncio==0.25.*
